+++ ImageEntity+CoreDataClass.swift
@@ -1,24 +1,16 @@
 //
-//  ImageData+CoreDataProperties.swift
+//  ImageEntity+CoreDataClass.swift
 //  Picterest
 //
-//  Created by Kai Kim on 2022/07/28.
+//  Created by Kai Kim on 2022/10/31.
 //
 //
 
 import Foundation
 import CoreData
 
-extension ImageEntity {
-
-    @nonobjc public class func fetchRequest() -> NSFetchRequest<ImageEntity> {
-        return NSFetchRequest<ImageEntity>(entityName: "ImageEntity")
-    }
-
-}
-
-extension ImageEntity : Identifiable {
-  
+@objc(ImageEntity)
+public class ImageEntity: NSManagedObject {
   func toDomain() -> Image {
     return .init(id: self.id,
                  imageURL: self.imageURL,
@@ -27,5 +19,11 @@ extension ImageEntity : Identifiable {
                  memo: self.memo,
                  isliked: self.isLiked)
   }
-  
 }
+
+//let id: String
+//let imageURL: URL
+//private(set) var width: Float
+//private(set) var height: Float
+//private(set) var memo: String?
+//private(set) var isLiked: Bool
