+++ Picterest/Domain/Usecases/FetchLikedImageUsecase.swift
@@ -1,0 +1,8 @@
+//
+//  FetchLikedImageUsecase.swift
+//  Picterest
+//
+//  Created by Kai Kim on 2022/10/12.
+//
+
+import Foundation
